Update README.md
This project prompts the user to select the starting and ending coordinates on the grid and then it finds the optimal path from source to destination using Breadth First Search algorithm (BFS).
If any obstacle is encountered while generating the path, it chooses an alternative route to provide the optimal path. We also added the colour coding to differentiate between free land, source and destination points, and the generated path.
If either the starting point or the destination point is an obstacle, the system will not generate any path.